+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/handlers/SplitQueryRequestHandler.java
@@ -5,6 +5,7 @@ package com.yahoo.bard.webservice.web.handlers;
 import static com.yahoo.bard.webservice.web.ErrorMessageFormat.EMPTY_INTERVAL_FORMAT;
 
 import com.yahoo.bard.webservice.application.MetricRegistryFactory;
+import com.yahoo.bard.webservice.druid.model.query.AllGranularity;
 import com.yahoo.bard.webservice.druid.model.query.DruidAggregationQuery;
 import com.yahoo.bard.webservice.druid.model.query.Granularity;
 import com.yahoo.bard.webservice.logging.RequestLog;
@@ -62,6 +63,11 @@ public class SplitQueryRequestHandler implements DataRequestHandler {
     ) {
         Granularity granularity = druidQuery.getGranularity();
         List<Interval> queryIntervals = druidQuery.getIntervals();
+
+        if (granularity instanceof AllGranularity) {
+           // For "all" grain there is only one response bucket, so can't really split
+           return next.handleRequest(context, request, druidQuery, response);
+        }
 
         Map<Interval, AtomicInteger> expectedIntervals = Collections.unmodifiableMap(
                 IntervalUtils.getSlicedIntervals(queryIntervals, granularity)